Update imports after renaming asset-store-controller-manager
