+++ components/cms-controller-manager/hack/crds/assetstore_v1alpha2_asset.yaml
@@ -1,0 +1,121 @@
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  creationTimestamp: null
+  labels:
+    controller-tools.k8s.io: "1.0"
+  name: assets.assetstore.kyma-project.io
+spec:
+  group: assetstore.kyma-project.io
+  names:
+    kind: Asset
+    plural: assets
+  scope: Namespaced
+  subresources:
+    status: {}
+  validation:
+    openAPIV3Schema:
+      properties:
+        apiVersion:
+          description: 'APIVersion defines the versioned schema of this representation
+            of an object. Servers should convert recognized schemas to the latest
+            internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources'
+          type: string
+        kind:
+          description: 'Kind is a string value representing the REST resource this
+            object represents. Servers may infer this from the endpoint the client
+            submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds'
+          type: string
+        metadata:
+          type: object
+        spec:
+          properties:
+            bucketRef:
+              properties:
+                name:
+                  type: string
+              required:
+              - name
+              type: object
+            source:
+              properties:
+                filter:
+                  type: string
+                mode:
+                  enum:
+                  - single
+                  - package
+                  - index
+                  type: string
+                mutationWebhookService:
+                  items:
+                    properties:
+                      endpoint:
+                        type: string
+                      metadata:
+                        type: object
+                      name:
+                        type: string
+                      namespace:
+                        type: string
+                    type: object
+                  type: array
+                url:
+                  type: string
+                validationWebhookService:
+                  items:
+                    properties:
+                      endpoint:
+                        type: string
+                      metadata:
+                        type: object
+                      name:
+                        type: string
+                      namespace:
+                        type: string
+                    type: object
+                  type: array
+              required:
+              - mode
+              - url
+              type: object
+          required:
+          - source
+          type: object
+        status:
+          properties:
+            assetRef:
+              properties:
+                assets:
+                  items:
+                    type: string
+                  type: array
+                baseUrl:
+                  type: string
+              required:
+              - baseUrl
+              type: object
+            lastHeartbeatTime:
+              format: date-time
+              type: string
+            message:
+              type: string
+            observedGeneration:
+              format: int64
+              type: integer
+            phase:
+              type: string
+            reason:
+              type: string
+          required:
+          - phase
+          - lastHeartbeatTime
+          - observedGeneration
+          type: object
+  version: v1alpha2
+status:
+  acceptedNames:
+    kind: ""
+    plural: ""
+  conditions: []
+  storedVersions: []
