+++ components/cms-controller-manager/pkg/apis/addtoscheme_assetstore_v1alpha2.go
@@ -1,7 +1,7 @@
 package apis
 
 import (
-	api "github.com/kyma-project/kyma/components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha2"
+	api "github.com/kyma-project/kyma/components/asset-store-controller-manager/pkg/apis/assetstore/v1alpha2"
 )
 
 func init() {
