+++ components/cms-controller-manager/pkg/controller/clusterdocstopic/clusterdocstopic_controller_suite_test.go
@@ -22,8 +22,7 @@ func TestMain(m *testing.M) {
 	t := &envtest.Environment{
 		CRDDirectoryPaths: []string{
 			filepath.Join("..", "..", "..", "config", "crds"),
-			// Hack for registering CRDs from vendored project
-			filepath.Join("..", "..", "..", "vendor", "github.com", "kyma-project", "kyma", "components", "assetstore-controller-manager", "config", "crds"),
+			filepath.Join("..", "..", "..", "hack", "crds"),
 		},
 	}
 	apis.AddToScheme(scheme.Scheme)
