+++ components/cms-controller-manager/pkg/controller/clusterdocstopic/clusterdocstopic_controller_test.go
@@ -2,7 +2,7 @@ package clusterdocstopic
 
 import (
 	"context"
-	"github.com/kyma-project/kyma/components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha2"
+	"github.com/kyma-project/kyma/components/asset-store-controller-manager/pkg/apis/assetstore/v1alpha2"
 	"github.com/kyma-project/kyma/components/cms-controller-manager/pkg/apis/cms/v1alpha1"
 	"github.com/kyma-project/kyma/components/cms-controller-manager/pkg/handler/docstopic/pretty"
 	"github.com/onsi/gomega"
