+++ components/cms-controller-manager/pkg/controller/docstopic/docstopic_controller.go
@@ -7,7 +7,7 @@ import (
 	"k8s.io/client-go/tools/record"
 	"time"
 
-	assetstoreapi "github.com/kyma-project/kyma/components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha2"
+	assetstoreapi "github.com/kyma-project/kyma/components/asset-store-controller-manager/pkg/apis/assetstore/v1alpha2"
 	cmsv1alpha1 "github.com/kyma-project/kyma/components/cms-controller-manager/pkg/apis/cms/v1alpha1"
 	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
